+++ egui_web/src/lib.rs
@@ -15,6 +15,8 @@
 #![warn(clippy::all, missing_crate_level_docs, rust_2018_idioms)]
 
 pub mod backend;
+#[cfg(feature = "use_glow_painter")]
+mod glow_wrapping;
 mod painter;
 pub mod screen_reader;
 pub mod webgl1;
@@ -26,6 +28,8 @@ use egui::mutex::Mutex;
 pub use wasm_bindgen;
 pub use web_sys;
 
+#[cfg(feature = "use_glow_painter")]
+use egui_glow;
 pub use painter::Painter;
 use std::cell::Cell;
 use std::rc::Rc;